Reset rajouté pour après mort Par contre `!!! Atention !
Le code est immonde besoin urgent d'être optimisé et rendu correct

+++ java/ch/epfl/cs107/play/game/areagame/Area.java
@@ -1,15 +1,12 @@
 package ch.epfl.cs107.play.game.areagame;
 
-import java.io.File;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import ch.epfl.cs107.play.game.Playable;
 import ch.epfl.cs107.play.game.actor.Actor;
 import ch.epfl.cs107.play.game.areagame.actor.Interactable;
 import ch.epfl.cs107.play.game.areagame.actor.Interactor;
+import ch.epfl.cs107.play.game.arpg.actor.ARPGMayor;
 import ch.epfl.cs107.play.game.arpg.actor.ARPGPlayer;
 import ch.epfl.cs107.play.io.FileSystem;
 import ch.epfl.cs107.play.math.DiscreteCoordinates;
@@ -45,34 +42,35 @@ public abstract class Area implements Playable {
     /// pause mechanics and menu to display. May be null
     /// - start indicate if area already begins, paused indicate if we display the pause menu
     private boolean started;//, paused;
-
-
+    private boolean onceReset = false;
+    private boolean onceReset2 = false;
 
     /** @return (float): camera scale factor, assume it is the same in x and y direction */
     public abstract float getCameraScaleFactor();
 
     /**
-     * Setter for the Behavior of this Area
-     * Please call this method in the begin method of every subclass
+     * Setter for the Behavior of this Area Please call this method in the begin method of every subclass
+     *
      * @param ab (AreaBehavior), not null
      */
-    protected final void setBehavior(AreaBehavior ab){
+    protected final void setBehavior(AreaBehavior ab) {
         this.areaBehavior = ab;
     }
 
     /**
      * Setter for the view Candidate
+     *
      * @param a (Actor), not null
      */
-    public final void setViewCandidate(Actor a){
+    public final void setViewCandidate(Actor a) {
         this.viewCandidate = a;
     }
-    
-    
+
+
     /**
-     * Add an actor to the actors list
-     * and to the behavior area cell if the actor is an Interactable
-     * and to the interactor list if the actor is an Interactor
+     * Add an actor to the actors list and to the behavior area cell if the actor is an Interactable and to the interactor list if the actor
+     * is an Interactor
+     *
      * @param a (Actor)(Interactor?)(Interactable?): the actor to add, not null
      * @param safeMode (Boolean): if True, the method ends
      */
@@ -80,13 +78,13 @@ public abstract class Area implements Playable {
 
         boolean errorHappen = false;
 
-        if(a instanceof Interactor)
-            errorHappen = !interactors.add((Interactor) a);
-        if(a instanceof Interactable)
+        if (a instanceof Interactor) { errorHappen = !interactors.add((Interactor) a); }
+        if (a instanceof Interactable) {
             errorHappen = errorHappen || !enterAreaCells(((Interactable) a), ((Interactable) a).getCurrentCells());
+        }
         errorHappen = errorHappen || !actors.add(a);
 
-        if(errorHappen && !safeMode) {
+        if (errorHappen && !safeMode) {
             System.out.println("Actor " + a + " cannot be completely added, so remove it from where it was");
             // Call it in safe mode to avoid recursive calls
             removeActor(a, true);
@@ -94,22 +92,22 @@ public abstract class Area implements Playable {
     }
 
     /**
-     * Remove an actor form the actor list
-     * and from the behavior area cell if the actor is an Interactable
-     * and from the interactor list if the actor is an Interactor
+     * Remove an actor form the actor list and from the behavior area cell if the actor is an Interactable and from the interactor list if
+     * the actor is an Interactor
+     *
      * @param a (Actor): the actor to remove, not null
      * @param safeMode (Boolean): if True, the method ends
      */
-    private void removeActor(Actor a, boolean safeMode){
+    private void removeActor(Actor a, boolean safeMode) {
         boolean errorHappen = false;
 
-        if(a instanceof Interactor)
-            errorHappen = !interactors.remove((Interactor)a);
-        if(a instanceof Interactable)
+        if (a instanceof Interactor) { errorHappen = !interactors.remove((Interactor) a); }
+        if (a instanceof Interactable) {
             errorHappen = errorHappen || !leaveAreaCells(((Interactable) a), ((Interactable) a).getCurrentCells());
+        }
         errorHappen = errorHappen || !actors.remove(a);
 
-        if(errorHappen && !safeMode) {
+        if (errorHappen && !safeMode) {
             System.out.println("Actor " + a + " cannot be completely removed, so add it from where it was");
             // Call it in safe mode to avoid recursive calls
             addActor(a, true);
@@ -118,81 +116,87 @@ public abstract class Area implements Playable {
 
     /**
      * Register an actor : will be added at next update
+     *
      * @param a (Actor): the actor to register, not null
      * @return (boolean): true if the actor is correctly registered
      */
-    public final boolean registerActor(Actor a){
+    public final boolean registerActor(Actor a) {
         //  finer Area strategies can be implemented here if wanted
         return registeredActors.add(a);
     }
 
     /**
      * Unregister an actor : will be removed at next update
+     *
      * @param a (Actor): the actor to unregister, not null
      * @return (boolean): true if the actor is correctly unregistered
      */
-    public final boolean unregisterActor(Actor a){
+    public final boolean unregisterActor(Actor a) {
         // finer Area strategies can be implemented here if wanted
         return unregisteredActors.add(a);
     }
 
     /**
      * Indicate if the given actor exists into the actor list
+     *
      * @param a (Actor): the given actor, may be null
      * @return (boolean): true if the given actor exists into actor list
      */
-    public boolean exists(Actor a){
+    public boolean exists(Actor a) {
         return actors.contains(a);
     }
 
 
     /**
      * Getter for the area width
+     *
      * @return (int) : the width in number of cols
      */
-    public int getWidth(){ 
+    public int getWidth() {
         return areaBehavior.getWidth();
     }
 
     /**
      * Getter for the area height
+     *
      * @return (int) : the height in number of rows
      */
-    public int getHeight(){ 
+    public int getHeight() {
         return areaBehavior.getHeight();
     }
 
     /** @return the Window Keyboard for inputs */
-    public final Keyboard getKeyboard () {
+    public final Keyboard getKeyboard() {
         return window.getKeyboard();
     }
-    
+
     /** @return the Window Mouse for inputs */
     public final Mouse getMouse() {
-    	return window.getMouse();
+        return window.getMouse();
     }
 
     /** @return the mouse position relatively to the area and the cells */
     public Vector getRelativeMousePosition() {
-    	return getMouse().getPosition();
+        return getMouse().getPosition();
     			/*.max(new Vector(0,0))
 				.min(new Vector(getWidth(),getHeight()));*/
     }
-    
+
     /** @return the mouse coordinates relatively to the area and the cells */
     public DiscreteCoordinates getRelativeMouseCoordinates() {
-    	Vector mousePosition = getRelativeMousePosition();
-    	DiscreteCoordinates mouseCoordinate = new DiscreteCoordinates((int)Math.floor(mousePosition.x), (int)Math.floor(mousePosition.y));
-    	return mouseCoordinate;
+        Vector mousePosition = getRelativeMousePosition();
+        DiscreteCoordinates mouseCoordinate = new DiscreteCoordinates((int) Math.floor(mousePosition.x), (int) Math.floor(mousePosition.y));
+        return mouseCoordinate;
     }
-    
-    /** @return (boolean): true if the method begin already called once. You can use resume() instead*/
+
+    /** @return (boolean): true if the method begin already called once. You can use resume() instead */
     public final boolean isStarted() {
         return started;
     }
 
     /**
      * If possible make the given interactable entity leave the given area cells
+     *
      * @param entity (Interactable), not null
      * @param coordinates (List of DiscreteCoordinates), may be empty but not null
      * @return (boolean): True if possible to leave
@@ -200,7 +204,7 @@ public abstract class Area implements Playable {
     public final boolean leaveAreaCells(Interactable entity, List<DiscreteCoordinates> coordinates) {
         // Finer Area strategies can be implemented here if wanted
         // Until now, the entity is put in a map waiting the update end to avoid concurrent exception during interaction
-        if(areaBehavior.canLeave(entity, coordinates)){
+        if (areaBehavior.canLeave(entity, coordinates)) {
             interactablesToLeave.put(entity, coordinates);
             return true;
         }
@@ -209,6 +213,7 @@ public abstract class Area implements Playable {
 
     /**
      * If possible make the given interactable entity enter the given area cells
+     *
      * @param entity (Interactable), not null
      * @param coordinates (List of DiscreteCoordinates), may be empty but not null
      * @return (boolean): True if possible to enter
@@ -216,15 +221,16 @@ public abstract class Area implements Playable {
     public final boolean enterAreaCells(Interactable entity, List<DiscreteCoordinates> coordinates) {
         // Finer Area strategies can be implemented here if wanted
         // Until now, the entity is put in a map waiting the update end to avoid concurrent exception during interaction
-        if(areaBehavior.canEnter(entity, coordinates)){
+        if (areaBehavior.canEnter(entity, coordinates)) {
             interactablesToEnter.put(entity, coordinates);
             return true;
         }
         return false;
     }
-    
+
     /**
      * Inform if the entity can enter the area cells
+     *
      * @param entity (Interactable), not null
      * @param coordinates (List of DiscreteCoordinates), may be empty but not null
      * @return (boolean): True if possible to enter
@@ -247,52 +253,65 @@ public abstract class Area implements Playable {
         interactablesToEnter = new HashMap<>();
         interactablesToLeave = new HashMap<>();
         viewCenter = Vector.ZERO;
-        //paused = false;
         started = true;
         return true;
     }
 
     /**
      * Resume method: Can be overridden
+     *
      * @param window (Window): display context, not null
      * @param fileSystem (FileSystem): given file system, not null
      * @return (boolean) : if the resume succeed, true by default
      */
-    public boolean resume(Window window, FileSystem fileSystem){
+    public boolean resume(Window window, FileSystem fileSystem) {
         return true;
     }
 
     @Override
     public void update(float deltaTime) {
-    	purgeRegistration();
 
-    	// Render actors
-    	for (Actor actor : actors) {
-    		actor.update(deltaTime);
-    	}
 
-    	// Realize interaction between interactors and their cells contents
-    	for (Interactor interactor : interactors) {
-    		if (interactor.wantsCellInteraction()) {
-    			areaBehavior.cellInteractionOf(interactor);
-    		}
-    		if (interactor.wantsViewInteraction()) {
-    			areaBehavior.viewInteractionOf(interactor);
-    		}
-    	}
+        purgeRegistration();
 
-    	// Update camera location
-    	updateCamera();
+        // Render actors
+        for (Actor actor : actors) {
+            actor.update(deltaTime);
+        }
 
-    	// Draw actors
-    	for (Actor actor : actors) {
-    		actor.bip(window);
-    		actor.draw(window);
-    	}
+        // Realize interaction between interactors and their cells contents
+        for (Interactor interactor : interactors) {
+            if (interactor.wantsCellInteraction()) {
+                areaBehavior.cellInteractionOf(interactor);
+            }
+            if (interactor.wantsViewInteraction()) {
+                areaBehavior.viewInteractionOf(interactor);
+            }
+        }
 
-     //   Keyboard keyboard = this.getKeyboard();
-        if(ARPGPlayer.reset && this.isStarted()){
-            //begin(this.window, fileSystem);
+        // Update camera location
+        updateCamera();
+
+        // Draw actors
+        for (Actor actor : actors) {
+            actor.bip(window);
+            actor.draw(window);
+        }
+        if (ARPGPlayer.isReset()) {
+            ARPGPlayer.setReset(false);
+            purgeRegistration();
+            onceReset = reset();
+            onceReset = true;
+            purgeRegistration();
+        }
+
+        if (onceReset) {
+            purgeRegistration();
+            onceReset = false;
+            reset();
+            purgeRegistration();
+
+
         }
 
     }
@@ -313,25 +332,25 @@ public abstract class Area implements Playable {
 
         // PART 2
         // - leave old cells
-        for (Map.Entry<Interactable, List<DiscreteCoordinates>> entry : interactablesToLeave.entrySet()){
+        for (Map.Entry<Interactable, List<DiscreteCoordinates>> entry : interactablesToLeave.entrySet()) {
             areaBehavior.leave(entry.getKey(), entry.getValue());
         }
         interactablesToLeave.clear();
         // - enter new cells
-        for (Map.Entry<Interactable, List<DiscreteCoordinates>> entry : interactablesToEnter.entrySet()){
+        for (Map.Entry<Interactable, List<DiscreteCoordinates>> entry : interactablesToEnter.entrySet()) {
             areaBehavior.enter(entry.getKey(), entry.getValue());
         }
         interactablesToEnter.clear();
     }
 
 
-    private void updateCamera () {
+    private void updateCamera() {
 
         // Update expected viewport center
         if (viewCandidate != null) {
             viewCenter = viewCandidate.getPosition();
-        }else { // Set default view to center
-        	viewCenter = new Vector(getWidth()/(float)2,getHeight()/(float)2);
+        } else { // Set default view to center
+            viewCenter = new Vector(getWidth() / (float) 2, getHeight() / (float) 2);
         }
         // Compute new viewport
         Transform viewTransform = Transform.I.scaled(getCameraScaleFactor()).translated(viewCenter);
@@ -341,15 +360,35 @@ public abstract class Area implements Playable {
     /**
      * Suspend method: Can be overridden, called before resume other
      */
-    public void suspend(){
+    public void suspend() {
         // Does nothing by default
     }
 
+    public boolean reset() {
+        boolean isResetOk = true;
+        Actor actor;
+        List<Actor> tmpActors = new LinkedList<>();
+
+        if (!onceReset) {
+            for (Actor i : actors) {
+                tmpActors.add(i);
+                isResetOk = unregisterActor(i) && isResetOk;
+            }
+            return isResetOk;
+
+        } else {
+
+            for (Actor i : tmpActors) {
+                isResetOk = registerActor(i) && isResetOk;
+            }
+            return true;
+        }
+    }
 
     @Override
     public void end() {
         // by default does nothing
-    	// can save the Area state somewhere if wanted
+        // can save the Area state somewhere if wanted
     }
 
 }
